requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
pulp>=2.7
pyyaml>=6.0
openpyxl>=3.1
requests>=2.28
scipy>=1.10
